Standardize on git_repo instead of git for fetching sources for all dependencies.
This follows Freedesktopsdk's standards.

(We will fix up platform and the rest of legacy in a future commit once we lock
in their public repositories.)

Of note, the 2nd parameter in the ref triplet is parsed as depth+1 (e.g. 0 here
equals depth of 1) by git_repo plugin.  This can greatly reduce the size of the
checkouts.  pytorch checkout shrinks from 4.2GB to 253MB as we are now just
checking out a shallow copy by default.