Tighten up the protobuf static linking as it was leaking and onnx erroring upon duplicated static registration in torchcodec and torch.
The protobuf-static library as-is right now does require that
dependencies take in absl shared libraries at run-time - so add
that as a run-time dependency.

Also bring in accelerate, cloudpickle, einops, farama-notifications,
gymnasium and psutil to support lerobot when using torch operators.

Back out the onnxruntime-shared-lib-python patch as we don't need it
and was causing an error in lerobot usecase.